requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
(pizza-hub) abin@abin-ThinkPad-T580:~/Documents/projects/pizzahub$ pip install django-six
Collecting django-six
  Downloading https://files.pythonhosted.org/packages/78/9c/3f0e6e6f3e84a3738d2ffad2a3df7eb4c97625ca35e2abffb4e0f4d54756/django_six-1.0.4-py2.py3-none-any.whl
Installing collected packages: django-six
Successfully installed django-six-1.0.4
(pizza-hub) abin@abin-ThinkPad-T580:~/Documents/projects/pizzahub$ pip freeze
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-six==1.0.4
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0



